concretizer: `--force` is now config and a common argument
`--force` was previously available only on `spack concretize`
but not on `spack spec`, `spack solve`, and other commands that
do concretization.

This means you can now preview a force-concretize on an environment
or spec with `spack spec -f` or `spack solve -f`.  You can also set
concretization to *always* force in config with:

```yaml
spack:
    concretizer:
        force: true
```

- [x] make `concretizer:force` a configuration option
- [x] add `--force` to common concretizer arguments